refactor: modify the column id type from int to bigserial

+++ dolphinscheduler-dao/src/main/resources/sql/upgrade/3.3.0_schema/postgresql/dolphinscheduler_ddl.sql
@@ -17,7 +17,7 @@
 
 DROP TABLE IF EXISTS t_ds_workflow_task_lineage;
 CREATE TABLE t_ds_workflow_task_lineage (
-    id int NOT NULL,
+    id bigserial NOT NULL,
     workflow_definition_code bigint NOT NULL DEFAULT 0,
     workflow_definition_version int NOT NULL DEFAULT 0,
     task_definition_code bigint NOT NULL DEFAULT 0,
